Lekce1/requirements.txt
flask
flask-cors
langchain
langchain-core
langchain-google-genai
google-generativeai
yfinance
openpyxl
requests
pydantic<2
